feat: enhance attachment previews and add file info display
- Updated DocumentAttachment, ImageAttachment, OtherAttachment, and TextAttachment components to include file information such as name, size, and extension.
- Adjusted layout and styling for better visual consistency across attachment types.
- Introduced a new MathBlock component for rendering LaTeX math expressions.
- Improved notification card styling and interaction feedback.
- Added an UpdateSection component for managing application updates, including version checking and downloading.
- Implemented virtual scrolling in TopicList for better performance with large topic lists.
- Registered a global event listener for sync completion to refresh assistant and topic data.

+++ src-tauri/src/vcp_modules/agent_types.rs
@@ -22,7 +22,7 @@ pub struct AgentConfig {
     pub model: String,
     /// 模型采样温度 (0.0-2.0)
     #[serde(default = "default_temperature")]
-    pub temperature: f32,
+    pub temperature: f64,
     /// 上下文 Token 限制
     #[serde(rename = "contextTokenLimit", default = "default_context_limit")]
     pub context_token_limit: i32,
@@ -52,7 +52,7 @@ fn default_agent_name() -> String {
 fn default_model() -> String {
     "gemini-2.5-flash".to_string()
 }
-fn default_temperature() -> f32 {
+fn default_temperature() -> f64 {
     1.0
 }
 fn default_context_limit() -> i32 {
